:art: Ajout du Fomulaire de candidature et suppression du fichier test
